convert contact page to modal, delete contact page
