--- src/app/contact/page.css
@@ -1,245 +1,0 @@
-.contact-page__main {
-  display: flex;
-  height: fit-content;
-  position: relative;
-  flex: 1;
-}
-
-.contact-page__contact {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  width: 100%;
-  padding: 40px;
-  box-shadow: inset black 5px 5px 10px;
-}
-
-.contact__border-image {
-  object-fit: cover;
-  height: auto;
-  width: 140px;
-}
-
-.contact-page__title {
-  align-self: center;
-  font-size: 50px;
-  color: #efe4d3;
-  letter-spacing: 3px;
-  margin: 0;
-  padding: 10px;
-  width: 100%;
-  max-width: 850px;
-  background-color: #191919;
-  text-align: center;
-  box-sizing: border-box;
-  border-top: black solid 2px;
-  border-left: black solid 2px;
-  border-right: black solid 2px;
-  border-radius: 8px 8px 0 0;
-}
-
-.form__wrapper {
-  position: relative;
-  box-sizing: border-box;
-  padding: 20px 60px;
-  width: 100%;
-  max-width: 850px;
-  display: flex;
-  flex-wrap: wrap;
-  flex-direction: column;
-  align-content: center;
-  background-color: #f2eee2;
-  text-align: center;
-  border-left: black solid 2px;
-  border-right: black solid 2px;
-  border-bottom: black solid 2px;
-  align-self: center;
-  border-radius: 0 0 8px 8px;
-}
-
-.form__title {
-  font-size: 32px;
-  line-height: 1.6rem;
-  margin: 5px;
-}
-
-.form__description {
-  line-height: 1.6rem;
-  font-weight: 400;
-  max-width: 500px;
-  margin: 0;
-  align-self: center;
-  font-size: 20px;
-}
-
-.form__description::after {
-  content: "";
-  display: block;
-  height: 0.5px;
-  width: 500px;
-  margin: 7px 0;
-  align-self: center;
-  border-bottom: black dotted 1.5px;
-}
-
-.form__accent {
-  position: absolute;
-  height: auto;
-  max-width: 90px;
-  width: 7vw;
-}
-
-.accent__top-left {
-  top: 22px;
-  left: 22px;
-}
-
-.accent__top-right {
-  top: 22px;
-  right: 22px;
-  transform: rotate(90deg);
-}
-
-.accent__bottom-left {
-  bottom: 22px;
-  left: 22px;
-  transform: rotate(270deg);
-}
-
-.accent__bottom-right {
-  bottom: 22px;
-  right: 22px;
-  transform: rotate(180deg);
-}
-
-@media (max-width: 1290px) {
-  .contact-page__contact {
-    justify-content: flex-start;
-    transform: translateY(0);
-  }
-
-  .contact-page__title {
-    font-size: 46px;
-    max-width: 710px;
-  }
-
-  .form__wrapper {
-    max-width: 710px;
-    padding: 20px 50px;
-  }
-
-  .form__title {
-    font-size: 28px;
-  }
-
-  .form__description {
-    line-height: 1.2rem;
-    max-width: 400px;
-    font-size: 18px;
-  }
-
-  .form__description::after {
-    width: 400px;
-  }
-}
-
-@media (max-width: 1070px) {
-  .contact-page__title {
-    font-size: 38px;
-    padding: 10px;
-  }
-
-  .form__title {
-    font-size: 26px;
-  }
-
-  .form__description {
-    font-size: 18px;
-  }
-
-  .contact__border-image {
-    width: 80px;
-  }
-}
-
-@media (max-width: 850px) {
-  .contact-page__contact {
-    border: none;
-  }
-
-  .contact__border-image {
-    display: none;
-  }
-
-  .form__wrapper {
-    padding: 10px 60px;
-  }
-
-  .contact-page__title {
-    font-size: 36px;
-    padding: 10px;
-  }
-}
-
-@media (max-width: 720px) {
-  .contact-page__contact {
-    padding: 40px 15px;
-  }
-
-  .form__wrapper {
-    max-width: 600px;
-    padding: 10px 30px;
-  }
-
-  .contact-page__title {
-    max-width: 600px;
-    font-size: 5vw;
-    padding: 12px;
-    font-size: 32px;
-  }
-
-  .form__title {
-    margin: 4px;
-  }
-
-  .form__description {
-    font-size: 16px;
-    margin-bottom: 5px;
-  }
-
-  .form__description::after {
-    margin: 10px 0;
-    width: 100%;
-  }
-}
-
-@media (max-width: 600px) {
-  .contact-page__contact {
-    padding: 15px;
-  }
-
-  .form__accent {
-    display: none;
-  }
-}
-
-@media (max-width: 450px) {
-  .form__wrapper {
-    padding: 10px 20px;
-  }
-
-  .form__title {
-    font-size: 5.5vw;
-    margin: 2px;
-  }
-
-  .form__description {
-    font-size: 15px;
-    line-height: 1.1rem;
-    margin-bottom: 5px;
-  }
-
-  .contact-form__input {
-    font-size: 14px;
-  }
-}
